+++ changelog.md
@@ -1,14 +1,17 @@
 # Changelog
 
-## [Fixes/Update] 7.31.2021 - 1
+## [Fixes/Update/Additions] 7.31.2021 - 1
 * Fixed issue with BU Lightsignal that could be without the actual signal
 * Fixed issue with crash on old worlds
 * Removed confusing off states
 * Removed unused states
-* Fixed issues with Changable Stage
+* Fixed issues with Changeable Stage
 * Updated internal rendering for name renderer
 * Fixed issue with height on RA signals
 * In GUI renderer now shows custom name plate
+* Added BUE2, BUE3 Signs
+* Added Station Number Sign
+* Added more ZS3 Textures
 
 ## [Fixed] 7.30.2021 - 1
 * Fixed issue with server client sync on redstone single mode